made deleting concepts shared code
