+++ react-client-app/src/components/DataTbl.jsx
@@ -50,7 +50,7 @@ import axios from 'axios'
 
 
 
-const DataTbl = () => {  
+const DataTbl = (props) => {  
     const value =parseInt(new URLSearchParams(window.location.search).get("search")) 
     //6284
     //21187
@@ -246,53 +246,7 @@ const DataTbl = () => {
 
 
     const handleDelete = (id, conceptId) => {
-        scanReportsRef.current = scanReportsRef.current.map((value) => value.id == id ? { ...value, conceptsLoaded: false } : value)
-        setScanReports(scanReportsRef.current)
-        //DEETE Request to API
-        useDelete(`scanreportconcepts/${conceptId}`)
-            .then(function (response) {
-                //Re-fetch the concepts for that particular field
-                getScanReportConcepts(id).then(scanreportconcepts => {
-                    if (scanreportconcepts.length > 0) {
-                        const conceptIds = scanreportconcepts.map(value => value.concept)
-                        useGet(`/omop/conceptsfilter/?concept_id__in=${conceptIds.join()}`)
-                            .then((values) => {
-                                scanreportconcepts = scanreportconcepts.map(element => ({ ...element, concept: values.find(con => con.concept_id == element.concept) }))
-                                scanReportsRef.current = scanReportsRef.current.map((value) => value.id == id ? { ...value, concepts: [...scanreportconcepts], conceptsLoaded: true } : value)
-                                setScanReports(scanReportsRef.current)
-                                setAlert({
-                                    status: 'success',
-                                    title: 'Concept Id Deleted',
-                                    description: 'Response: ' + response.status + ' ' + response.statusText
-                                })
-                                onOpen()
-                            });
-                    }
-                    else {
-                        scanReportsRef.current = scanReportsRef.current.map((value) => value.id == id ? { ...value, concepts: [], conceptsLoaded: true } : value)
-                        setScanReports(scanReportsRef.current)
-                        setAlert({
-                            status: 'success',
-                            title: 'Concept Id Deleted',
-                            description: 'Response: ' + response.status + ' ' + response.statusText
-                        })
-                        onOpen()
-                    }
-                })
-            })
-            .catch(function (error) {
-                scanReportsRef.current = scanReportsRef.current.map((value) => value.id == id ? { ...value, conceptsLoaded: true } : value)
-                setScanReports(scanReportsRef.current)
-                if (typeof (error) !== 'undefined' && error.response != null) {
-                    setAlert({
-                        status: 'error',
-                        title: 'Unable to delete Concept id from value',
-                        description: 'Response: ' + error.response.status + ' ' + error.response.statusText
-                    })
-                    onOpen()
-
-                }
-            })
+        props.handleDelete(id,conceptId,scanReportsRef,setScanReports,setAlert,onOpen)
     } 
 
     if (error){
